pyproject.toml
[project]
name = "health-ai-benchmarking"
version = "0.1.0"
description = "A benchmarking framework for evaluating LLM accuracy and safety in suicide risk assessment using the C-SSRS scale."
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "openai>=2.29.0",
    "google-genai>=0.1.0",
    "anthropic>=0.25.0",
    "python-dotenv>=1.0.1",
    "requests>=2.31.0",
]

[tool.uv]
managed = true
